fix(web3): extract decode-calldata core to fix workflow bundler
The workflow bundler breaks when functions are exported from files
containing "use step" directives — it pulls all transitive deps
(ethers, drizzle, @solana/web3.js, nanoid) into the restricted
workflow runtime.

Fix: extract core decode logic into decode-calldata-core.ts (no
"use step"), which both decode-calldata and assess-risk import.
Also replace AI SDK with direct fetch to Anthropic API since the
SDK has Node.js-only transitive deps incompatible with workflow
runtime.
